pyproject.toml
[tool.poetry]
name = "thread-maker-discord-bot"
version = "0.2.0"
description = ""
authors = ["vitivalkoinen <if.snow.else.neige@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
